room system: createRoom moved from factory to own file
